docqa/docqa_api/utils/pdf_utils.py
import fitz  # PyMuPDF


def extract_text_from_pdf(file_path: str) -> str:
    pdf = fitz.open(file_path)
    text = ""
    for page in pdf:
        text += page.get_text("text") + "\n"
    return text.strip()
